requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.10
sqlparse==0.4.4
typing-extensions==4.9.0
python-decouple==3.5
drf-spectacular==0.18.0
psycopg2-binary==2.9.1
pytest
factory_boy